requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
requests>=2.28.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
